requirements.txt
gradio
python-dotenv
chromadb
requests
fastapi
uvicorn[standard]
pydantic
modal
Pillow
pytest
pytest-cov